requirements.txt
fastapi[all]==0.115.5
uvicorn[standard]==0.32.0
pandas== 2.2.3
scikit-learn==1.5.2
joblib==1.4.2
langchain==1.1.0
langchain-google-genai==3.2.0
langchain-community>=0.4.1
psycopg2-binary==2.9.11
sqlmodel==0.0.27
psycopg2-binary==2.9.11
python-dotenv==1.0.1